pyproject.toml
[tool.poetry]
name = "az-snowflake-broker"
version = "0.1.0"
description = "Communicate between azure cli and snowflake to manage azure resources and store azure information inside Snowflake"
authors = ["sebastian hansen <sebastian.hansen@dataalchemy.dev>"]

[tool.poetry.dependencies]
python ="==3.8.*"
pyOpenSSL = ">=22.0.0"
pyarrow = "8.0.0"
snowflake-snowpark-python = "^1.0.0"
snowflake-connector-python = "^2.9.0"
azure-cli = "^2.43.0"
python-dotenv = "^0.21.0"
tqdm = "^4.64.1"
pandas = "^1.5.2"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
